(chore) move LogEntry to its own file and use optional types instead of Maybe for props
